+++ backend/canisters/group/impl/src/updates/change_role.rs
@@ -9,7 +9,7 @@ use group_community_common::ExpiringMember;
 use types::{CanisterId, GroupRole, UserId};
 use user_index_canister_c2c_client::{lookup_user, LookupUserError};
 
-#[update(candid = true, msgpack = true)]
+#[update(msgpack = true)]
 #[trace]
 async fn change_role(args: Args) -> Response {
     run_regular_jobs();