+++ backend/canisters/group/impl/src/updates/claim_prize.rs
@@ -8,7 +8,7 @@ use group_canister::claim_prize::{Response::*, *};
 use ledger_utils::{create_pending_transaction, process_transaction};
 use types::{CanisterId, CompletedCryptoTransaction, PendingCryptoTransaction, UserId};
 
-#[update(candid = true, msgpack = true)]
+#[update(msgpack = true)]
 #[trace]
 async fn claim_prize(args: Args) -> Response {
     run_regular_jobs();