+++ backend/canisters/community/impl/src/updates/create_channel.rs
@@ -17,7 +17,7 @@ use utils::text_validation::{
     validate_description, validate_group_name, validate_rules, NameValidationError, RulesValidationError,
 };
 
-#[update(candid = true, msgpack = true)]
+#[update(msgpack = true)]
 #[trace]
 fn create_channel(args: Args) -> Response {
     run_regular_jobs();